Started working on improvements suggested in PR
