+++ CODEOWNERS
@@ -70,7 +70,7 @@ homeassistant/components/blueprint/* @home-assistant/core
 homeassistant/components/bmp280/* @belidzs
 homeassistant/components/bmw_connected_drive/* @gerard33 @rikroe
 homeassistant/components/bond/* @prystupa
-homeassistant/components/boschshc/* @tschamm @cabrust
+homeassistant/components/bosch_shc/* @tschamm @cabrust
 homeassistant/components/braviatv/* @bieniu
 homeassistant/components/broadlink/* @danielhiversen @felipediel
 homeassistant/components/brother/* @bieniu
