+++ homeassistant/components/bosch_shc/binary_sensor.py
@@ -1,0 +1,54 @@
+"""Platform for binarysensor integration."""
+import logging
+
+from boschshcpy import SHCSession, SHCShutterContact
+
+from homeassistant.components.binary_sensor import (
+    DEVICE_CLASS_DOOR,
+    DEVICE_CLASS_WINDOW,
+    BinarySensorEntity,
+)
+from homeassistant.const import CONF_IP_ADDRESS
+
+from .const import DOMAIN
+from .entity import SHCEntity
+
+_LOGGER = logging.getLogger(__name__)
+
+
+async def async_setup_entry(hass, config_entry, async_add_entities):
+    """Set up the SHC binary sensor platform."""
+    entities = []
+    session: SHCSession = hass.data[DOMAIN][config_entry.entry_id]
+
+    for binarysensor in session.device_helper.shutter_contacts:
+        entities.append(
+            ShutterContactSensor(
+                device=binarysensor,
+                room_name=session.room(binarysensor.room_id).name,
+                controller_ip=config_entry.data[CONF_IP_ADDRESS],
+            )
+        )
+
+    if entities:
+        async_add_entities(entities)
+
+
+class ShutterContactSensor(SHCEntity, BinarySensorEntity):
+    """Representation of a SHC shutter contact sensor."""
+
+    @property
+    def is_on(self):
+        """Return the state of the sensor."""
+        return self._device.state == SHCShutterContact.ShutterContactService.State.OPEN
+
+    @property
+    def device_class(self):
+        """Return the class of this device, from component DEVICE_CLASSES."""
+        switcher = {
+            SHCShutterContact.DeviceClass.ENTRANCE_DOOR: DEVICE_CLASS_DOOR,
+            SHCShutterContact.DeviceClass.REGULAR_WINDOW: DEVICE_CLASS_WINDOW,
+            SHCShutterContact.DeviceClass.FRENCH_WINDOW: DEVICE_CLASS_DOOR,
+            SHCShutterContact.DeviceClass.GENERIC: DEVICE_CLASS_WINDOW,
+        }
+        return switcher.get(self._device.device_class, DEVICE_CLASS_WINDOW)
