+++ homeassistant/components/bosch_shc/const.py
@@ -1,10 +1,6 @@
 """Constants for the Bosch SHC integration."""
 
-ATTR_NAME = "name"
-
 CONF_SSL_CERTIFICATE = "ssl_certificate"
 CONF_SSL_KEY = "ssl_key"
 
-DOMAIN = "boschshc"
-
-SERVICE_TRIGGER_SCENARIO = "trigger_scenario"
+DOMAIN = "bosch_shc"
