+++ homeassistant/components/bosch_shc/manifest.json
@@ -1,0 +1,11 @@
+{
+  "domain": "bosch_shc",
+  "name": "Bosch Smart Home Controller Interface API",
+  "config_flow": true,
+  "documentation": "https://www.home-assistant.io/integrations/bosch_shc",
+  "requirements": ["boschshcpy==0.1.8"],
+  "codeowners": [
+    "@tschamm",
+    "@cabrust"
+  ]
+}
