+++ homeassistant/components/bosch_shc/strings.json
