+++ homeassistant/components/bosch_shc/translations/de.json
