+++ homeassistant/components/bosch_shc/translations/en.json
