--- homeassistant/components/boschshc/manifest.json
@@ -1,15 +1,0 @@
-{
-  "domain": "boschshc",
-  "name": "Bosch Smart Home Controller Interface API",
-  "config_flow": true,
-  "documentation": "https://www.home-assistant.io/integrations/boschshc",
-  "requirements": ["boschshcpy==0.1.4"],
-  "ssdp": [],
-  "zeroconf": [],
-  "homekit": {},
-  "dependencies": [],
-  "codeowners": [
-    "@tschamm",
-    "@cabrust"
-  ]
-}
