--- homeassistant/components/boschshc/services.yaml
@@ -1,13 +1,0 @@
-smokedetector_check:
-  description: Calls test routine of Bosch Smoke Detector.
-  fields:
-    entity_id:
-      description: Name(s) of the entities to set
-      example: "binary_sensor.smokedetector_living_room"
-
-trigger_scenario:
-  description: Triggers a scenario on the Bosch SHC.
-  fields:
-    name:
-      description: Name of the scenario to trigger
-      example: "Scenario Name"
