+++ requirements_all.txt
@@ -382,8 +382,8 @@ blockchain==1.4.4
 # homeassistant.components.bond
 bond-api==0.1.12
 
-# homeassistant.components.boschshc
-boschshcpy==0.1.4
+# homeassistant.components.bosch_shc
+boschshcpy==0.1.8
 
 # homeassistant.components.amazon_polly
 # homeassistant.components.route53
